+++ muyie-starter-oss/src/main/java/com/muyie/oss/controller/OssUploadController.java
@@ -8,8 +8,8 @@ import com.muyie.exception.ErrorCodeDefaults;
 import com.muyie.exception.ExceptionUtil;
 import com.muyie.oss.autoconfigure.OssProperties;
 import com.muyie.oss.context.OssKeyGenerator;
-import com.muyie.oss.model.BucketProfile;
-import com.muyie.oss.model.StoreResult;
+import com.muyie.oss.model.StorageConfig;
+import com.muyie.oss.model.StorageInfo;
 import com.muyie.oss.service.OssService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -46,49 +46,41 @@ public class OssUploadController {
   /**
    * 单文件上传 - 注意要添加请求头：content-type=multipart/form-data
    *
-   * @param bucketKey 指定上传的BucketKey
-   * @param file      指定上传的文件
-   * @param prefix    指定上传的目录
+   * @param key  指定上传的BucketKey
+   * @param file 指定上传的文件
    * @return 返回上传结果信息
    */
-  @PostMapping("single")
+  @PostMapping(value = "single", consumes = "multipart/form-data")
   @Operation(summary = "单文件上传")
-  public MultiResponse<StoreResult> uploadSingle(@RequestParam("bucketKey") String bucketKey,
-                                                 @RequestParam("file") MultipartFile file,
-                                                 @RequestParam(value = "prefix", required = false) String prefix) {
-    return this.uploadMultipart(bucketKey, new MultipartFile[]{file}, prefix);
+  public MultiResponse<StorageInfo> uploadSingle(@RequestParam("key") String key, @RequestParam("file") MultipartFile file) {
+    return this.uploadMultipart(key, new MultipartFile[]{file});
   }
 
   /**
    * 多文件上传 - 注意要添加请求头：content-type=multipart/form-data
    *
-   * @param bucketKey 指定上传的BucketKey
-   * @param files     指定上传的文件列表
-   * @param prefix    指定上传的目录
+   * @param key   指定上传的BucketKey
+   * @param files 指定上传的文件列表
    * @return 返回上传结果信息
    */
-  @PostMapping("multipart")
+  @PostMapping(value = "multipart", consumes = "multipart/form-data")
   @Operation(summary = "多文件上传")
-  public MultiResponse<StoreResult> uploadMultipart(@RequestParam("bucketKey") String bucketKey,
-                                                    @RequestParam("files") MultipartFile[] files,
-                                                    @RequestParam(value = "prefix", required = false) String prefix) {
+  public MultiResponse<StorageInfo> uploadMultipart(@RequestParam("key") String key, @RequestParam("files") MultipartFile[] files) {
     try {
       if (files == null) {
         throw ExceptionUtil.business(ErrorCodeDefaults.A0700);
       }
-      BucketProfile bucketProfile = ossProperties.getBucketProfile(bucketKey);
-      this.assertMaxFiles(bucketProfile, files.length);
-      String targetPrefix = this.getPrefix(bucketProfile, prefix);
-      this.anyMatchPrefix(bucketProfile, targetPrefix);
-      List<StoreResult> list = Lists.newArrayList();
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      this.assertMaxFiles(config, files.length);
+      List<StorageInfo> list = Lists.newArrayList();
       for (MultipartFile file : files) {
         ExceptionUtil.validate().rewrite("文件不能为空").doThrow(file.isEmpty());
         String fileName = file.getOriginalFilename();
         log.info("上传的源文件名为：" + fileName);
-        String objectKey = ossKeyGenerator.getObjectKey(targetPrefix, file);
-        this.anyMatchSuffix(bucketProfile, objectKey);
-        StoreResult storeResult = ossService.putObject(bucketKey, objectKey, file.getInputStream());
-        list.add(storeResult);
+        String objectKey = ossKeyGenerator.getObjectKey(config.getFolder(), file);
+        this.anyMatchSuffix(config, objectKey);
+        StorageInfo storageInfo = ossService.putObject(key, objectKey, file.getInputStream());
+        list.add(storageInfo);
       }
       return MultiResponse.of(list);
     } catch (BaseException e) {
@@ -99,31 +91,25 @@ public class OssUploadController {
   }
 
   /**
-   * 上传网络流 - 注意要添加请求头：content-type=multipart/form-data
+   * 上传网络流
    *
-   * @param bucketKey 指定上传的BucketKey
-   * @param url       网络资源（主要是图片）
-   * @param prefix    指定上传的目录
-   * @param suffix    指定上传的后缀名
+   * @param key    指定上传的BucketKey
+   * @param url    网络资源（主要是图片）
+   * @param suffix 指定上传的后缀名
    * @return 返回上传结果信息
    */
   @PostMapping("url")
   @Operation(summary = "上传网络流")
-  public SingleResponse<StoreResult> uploadUrl(@RequestParam("bucketKey") String bucketKey,
-                                               @RequestParam("url") String url,
-                                               @RequestParam(value = "prefix", required = false) String prefix,
-                                               @RequestParam(value = "suffix", required = false) String suffix) {
+  public SingleResponse<StorageInfo> uploadUrl(@RequestParam("key") String key, @RequestParam("url") String url, @RequestParam(value = "suffix", required = false) String suffix) {
     try {
-      BucketProfile bucketProfile = ossProperties.getBucketProfile(bucketKey);
-      String targetPrefix = this.getPrefix(bucketProfile, prefix);
-      this.anyMatchPrefix(bucketProfile, targetPrefix);
+      StorageConfig config = ossProperties.getStorageConfig(key);
       if (StringUtils.isBlank(suffix)) {
         suffix = StringUtils.substringAfterLast(url, ".");
       }
-      String objectKey = ossKeyGenerator.getObjectKey(targetPrefix, suffix);
-      this.anyMatchSuffix(bucketProfile, objectKey);
-      StoreResult storeResult = ossService.putObject(bucketKey, objectKey, url);
-      return SingleResponse.of(storeResult);
+      String objectKey = ossKeyGenerator.getObjectKey(config.getFolder(), suffix);
+      this.anyMatchSuffix(config, objectKey);
+      StorageInfo storageInfo = ossService.putObject(key, objectKey, url);
+      return SingleResponse.of(storageInfo);
     } catch (BaseException e) {
       throw e;
     } catch (Exception e) {
@@ -132,29 +118,23 @@ public class OssUploadController {
   }
 
   /**
-   * 上传字节流 - 注意要添加请求头：content-type=multipart/form-data
+   * 上传字节流
    *
-   * @param bucketKey 指定上传的BucketKey
-   * @param object    上传字节流（Base64编码的字节对象）
-   * @param prefix    指定上传的目录
-   * @param suffix    指定上传的后缀名
+   * @param key    指定上传的BucketKey
+   * @param object 上传字节流（Base64编码的字节对象）
+   * @param suffix 指定上传的后缀名
    * @return 返回上传结果信息
    */
   @PostMapping("object")
   @Operation(summary = "上传字节流")
-  public SingleResponse<StoreResult> uploadObject(@RequestParam("bucketKey") String bucketKey,
-                                                  @RequestParam("object") String object,
-                                                  @RequestParam(value = "prefix", required = false) String prefix,
-                                                  @RequestParam(value = "suffix") String suffix) {
+  public SingleResponse<StorageInfo> uploadObject(@RequestParam("key") String key, @RequestParam("object") String object, @RequestParam(value = "suffix") String suffix) {
     try {
-      BucketProfile bucketProfile = ossProperties.getBucketProfile(bucketKey);
-      String targetPrefix = this.getPrefix(bucketProfile, prefix);
-      this.anyMatchPrefix(bucketProfile, targetPrefix);
-      String objectKey = ossKeyGenerator.getObjectKey(targetPrefix, suffix);
-      this.anyMatchSuffix(bucketProfile, objectKey);
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      String objectKey = ossKeyGenerator.getObjectKey(config.getFolder(), suffix);
+      this.anyMatchSuffix(config, objectKey);
       byte[] bytes = Base64Utils.decodeFromUrlSafeString(object);
-      StoreResult storeResult = ossService.putObject(bucketKey, objectKey, bytes);
-      return SingleResponse.of(storeResult);
+      StorageInfo storageInfo = ossService.putObject(key, objectKey, bytes);
+      return SingleResponse.of(storageInfo);
     } catch (BaseException e) {
       throw e;
     } catch (Exception e) {
@@ -165,46 +145,22 @@ public class OssUploadController {
   /**
    * 校验文件上传指定的文件后缀名称是否匹配
    *
-   * @param bucketProfile Bucket配置信息
-   * @param objectKey     对象名称
+   * @param config    对象存储配置
+   * @param objectKey 对象名称
    */
-  private void anyMatchSuffix(BucketProfile bucketProfile, String objectKey) {
-    boolean anyMatch = Arrays.stream(bucketProfile.getSuffixSupports()).anyMatch(s -> StringUtils.endsWithIgnoreCase(objectKey, s));
+  private void anyMatchSuffix(StorageConfig config, String objectKey) {
+    boolean anyMatch = Arrays.stream(config.getSuffixSupports()).anyMatch(s -> StringUtils.endsWithIgnoreCase(objectKey, s));
     ExceptionUtil.business(ErrorCodeDefaults.A0701).doThrow(!anyMatch);
-  }
-
-  /**
-   * 校验文件上传指定的目录是否匹配
-   *
-   * @param bucketProfile Bucket配置信息
-   * @param prefix        指定文件上传的目录
-   */
-  private void anyMatchPrefix(BucketProfile bucketProfile, String prefix) {
-    boolean anyMatch = Arrays.stream(bucketProfile.getPrefixSupports()).anyMatch(s -> StringUtils.endsWithIgnoreCase(prefix, s));
-    ExceptionUtil.business(ErrorCodeDefaults.A0701).rewrite("文件上传目录不支持").doThrow(!anyMatch);
   }
 
   /**
    * 校验用户上传的文件数量是否超限
    *
-   * @param bucketProfile Bucket配置信息
-   * @param size          上传的文件数量
+   * @param config 对象存储配置
+   * @param size   上传的文件数量
    */
-  private void assertMaxFiles(BucketProfile bucketProfile, int size) {
-    ExceptionUtil.business(ErrorCodeDefaults.A0425, "文件上传数量超限：size=" + size).doThrow(size > bucketProfile.getMaxFiles());
-  }
-
-  /**
-   * 获取文件上传的目录
-   *
-   * @param bucketProfile Bucket配置信息
-   * @param prefix        指定文件上传的目录
-   */
-  private String getPrefix(BucketProfile bucketProfile, String prefix) {
-    if (bucketProfile.isAllowPrefix() && StringUtils.isNotBlank(prefix)) {
-      return prefix;
-    }
-    return bucketProfile.getDefaultPrefix();
+  private void assertMaxFiles(StorageConfig config, int size) {
+    ExceptionUtil.business(ErrorCodeDefaults.A0425, "文件上传数量超限：size=" + size).doThrow(size > config.getMaxFiles());
   }
 
 }