+++ muyie-starter-oss/src/main/java/com/muyie/oss/service/OssService.java
@@ -4,8 +4,8 @@ import com.aliyun.oss.OSS;
 import com.aliyun.oss.model.OSSObject;
 import com.aliyun.oss.model.ObjectMetadata;
 import com.aliyun.oss.model.PutObjectRequest;
-import com.muyie.oss.model.BucketProfile;
-import com.muyie.oss.model.StoreResult;
+import com.muyie.oss.model.StorageConfig;
+import com.muyie.oss.model.StorageInfo;
 
 import java.io.File;
 import java.io.InputStream;
@@ -26,142 +26,142 @@ public interface OssService {
   OSS getOssClient();
 
   /**
-   * 获取Bucket配置信息
+   * 获取 Bucket 配置信息
    *
-   * @param bucketKey key
+   * @param key 配置StorageConfig的key
    * @return 配置信息
    */
-  BucketProfile getBucketProfile(String bucketKey);
+  StorageConfig getStorageConfig(String key);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey   配置BucketProfile的key
-   * @param objectKey   文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key         配置StorageConfig的key
+   * @param objectKey   文件存储路径
    * @param inputStream 文件对象
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, InputStream inputStream);
+  StorageInfo putObject(String key, String objectKey, InputStream inputStream);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey   配置BucketProfile的key
-   * @param objectKey   文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key         配置StorageConfig的key
+   * @param objectKey   文件存储路径
    * @param inputStream 文件对象
    * @param metadata    对象元信息
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, InputStream inputStream, ObjectMetadata metadata);
+  StorageInfo putObject(String key, String objectKey, InputStream inputStream, ObjectMetadata metadata);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @param file      文件对象
    * @param metadata  对象元信息
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, File file, ObjectMetadata metadata);
+  StorageInfo putObject(String key, String objectKey, File file, ObjectMetadata metadata);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @param file      文件对象
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, File file);
+  StorageInfo putObject(String key, String objectKey, File file);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @param bytes     文件对象
    * @param metadata  对象元信息
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, byte[] bytes, ObjectMetadata metadata);
+  StorageInfo putObject(String key, String objectKey, byte[] bytes, ObjectMetadata metadata);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @param bytes     文件对象
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, byte[] bytes);
+  StorageInfo putObject(String key, String objectKey, byte[] bytes);
 
   /**
    * 文件对象上传（上传网络流）
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @param url       资源网址
    * @param metadata  对象元信息
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, String url, ObjectMetadata metadata);
+  StorageInfo putObject(String key, String objectKey, String url, ObjectMetadata metadata);
 
   /**
    * 文件对象上传（上传网络流）
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @param url       资源网址
-   * @return 上传结果
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, String objectKey, String url);
+  StorageInfo putObject(String key, String objectKey, String url);
 
   /**
    * 文件对象上传
    *
-   * @param bucketKey        配置BucketProfile的key
-   * @param putObjectRequest 原生的文件上传对象（bucketName通过bucketKey自动获取）
-   * @return 上传结果
+   * @param key              配置StorageConfig的key
+   * @param putObjectRequest 原生的文件上传对象（参数bucketName将通过key自动获取）
+   * @return 对象存储信息
    */
-  StoreResult putObject(String bucketKey, PutObjectRequest putObjectRequest);
+  StorageInfo putObject(String key, PutObjectRequest putObjectRequest);
 
   /**
    * 复制文件
    *
-   * @param fromBucketKey 源文件配置BucketProfile的key
+   * @param fromKey       源文件配置StorageConfig的key
    * @param fromObjectKey 源文件存储路径
-   * @param toBucketKey   目标文件配置BucketProfile的key
+   * @param toKey         目标文件配置StorageConfig的key
    * @param toObjectKey   目标文件存储路径
    * @param isDelete      是否删除源文件
    * @return 复制后的文件存储信息
    */
-  StoreResult copyObject(String fromBucketKey, String fromObjectKey, String toBucketKey, String toObjectKey, boolean isDelete);
+  StorageInfo copyObject(String fromKey, String fromObjectKey, String toKey, String toObjectKey, boolean isDelete);
 
   /**
    * 获取文件
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @return 结果
    */
-  OSSObject getObject(String bucketKey, String objectKey);
+  OSSObject getObject(String key, String objectKey);
 
   /**
    * 删除文件
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    */
-  void deleteObject(String bucketKey, String objectKey);
+  void deleteObject(String key, String objectKey);
 
   /**
    * 获取对象元信息
    *
-   * @param bucketKey 配置BucketProfile的key
-   * @param objectKey 文件存储路径，e.g. temp/v1-2390293382.png
+   * @param key       配置StorageConfig的key
+   * @param objectKey 文件存储路径
    * @return 结果
    */
-  ObjectMetadata getObjectMetadata(String bucketKey, String objectKey);
+  ObjectMetadata getObjectMetadata(String key, String objectKey);
 
 }