+++ muyie-starter-oss/src/main/java/com/muyie/oss/service/impl/OssServiceImpl.java
@@ -7,8 +7,8 @@ import com.muyie.exception.ErrorCodeDefaults;
 import com.muyie.exception.ExceptionUtil;
 import com.muyie.oss.autoconfigure.OssProperties;
 import com.muyie.oss.context.OssUploadCallback;
-import com.muyie.oss.model.BucketProfile;
-import com.muyie.oss.model.StoreResult;
+import com.muyie.oss.model.StorageConfig;
+import com.muyie.oss.model.StorageInfo;
 import com.muyie.oss.service.OssService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +21,7 @@ import java.io.InputStream;
 import java.net.URL;
 
 /**
- * 简单文件操作实现
+ * 简单文件操作实现（上传、下载、删除、复制）
  *
  * @author larry.qi
  * @since 2.7.14
@@ -40,174 +40,174 @@ public class OssServiceImpl implements OssService {
   }
 
   @Override
-  public BucketProfile getBucketProfile(String bucketKey) {
-    return ossProperties.getBucketProfile(bucketKey);
+  public StorageConfig getStorageConfig(String key) {
+    return ossProperties.getStorageConfig(key);
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, InputStream inputStream) {
+  public StorageInfo putObject(String key, String objectKey, InputStream inputStream) {
     try {
-      Assert.notNull(bucketKey, "bucketKey must be not null");
+      Assert.notNull(key, "key must be not null");
       Assert.notNull(objectKey, "objectKey must be not null");
       Assert.notNull(inputStream, "inputStream must be not null");
-      BucketProfile bp = ossProperties.getBucketProfile(bucketKey);
-      log.info("putObject bucketKey={}, bucketName={}", bucketKey, bp.getBucket());
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      log.info("putObject key={}, bucketName={}", key, config.getBucket());
       log.info("putObject objectKey={}", objectKey);
-      PutObjectResult result = ossClient.putObject(bp.getBucket(), objectKey, inputStream);
+      PutObjectResult result = ossClient.putObject(config.getBucket(), objectKey, inputStream);
       Assert.notNull(result, "PutObjectResult must be not null");
       log.info("PutObjectResult {}", JSON.toJSONString(result));
-      StoreResult storeResult = new StoreResult();
-      storeResult.setBucket(bp.getBucket());
-      storeResult.setObjectKey(objectKey);
-      storeResult.setObjectUrl(bp.getBaseUrl() + objectKey);
-      storeResult.setEtag(result.getETag());
-      ossUploadCallback.callback(storeResult);
-      return storeResult;
+      StorageInfo storageInfo = new StorageInfo();
+      storageInfo.setBucket(config.getBucket());
+      storageInfo.setObjectKey(objectKey);
+      storageInfo.setObjectUrl(config.getBaseUrl() + objectKey);
+      storageInfo.setEtag(result.getETag());
+      ossUploadCallback.callback(storageInfo);
+      return storageInfo;
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, InputStream inputStream, ObjectMetadata metadata) {
+  public StorageInfo putObject(String key, String objectKey, InputStream inputStream, ObjectMetadata metadata) {
     try {
-      Assert.notNull(bucketKey, "bucketKey must be not null");
+      Assert.notNull(key, "key must be not null");
       Assert.notNull(objectKey, "objectKey must be not null");
       Assert.notNull(inputStream, "inputStream must be not null");
       Assert.notNull(metadata, "metadata must be not null");
-      BucketProfile bp = ossProperties.getBucketProfile(bucketKey);
-      log.info("putObject bucketKey={}, bucketName={}", bucketKey, bp.getBucket());
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      log.info("putObject key={}, bucketName={}", key, config.getBucket());
       log.info("putObject objectKey={}", objectKey);
-      PutObjectResult result = ossClient.putObject(bp.getBucket(), objectKey, inputStream, metadata);
+      PutObjectResult result = ossClient.putObject(config.getBucket(), objectKey, inputStream, metadata);
       Assert.notNull(result, "PutObjectResult must be not null");
       log.info("PutObjectResult {}", JSON.toJSONString(result));
-      StoreResult storeResult = new StoreResult();
-      storeResult.setBucket(bp.getBucket());
-      storeResult.setObjectKey(objectKey);
-      storeResult.setObjectUrl(bp.getBaseUrl() + objectKey);
-      storeResult.setEtag(result.getETag());
-      ossUploadCallback.callback(storeResult);
-      return storeResult;
+      StorageInfo storageInfo = new StorageInfo();
+      storageInfo.setBucket(config.getBucket());
+      storageInfo.setObjectKey(objectKey);
+      storageInfo.setObjectUrl(config.getBaseUrl() + objectKey);
+      storageInfo.setEtag(result.getETag());
+      ossUploadCallback.callback(storageInfo);
+      return storageInfo;
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, File file, ObjectMetadata metadata) {
+  public StorageInfo putObject(String key, String objectKey, File file, ObjectMetadata metadata) {
     try (FileInputStream inputStream = new FileInputStream(file)) {
-      return this.putObject(bucketKey, objectKey, inputStream, metadata);
+      return this.putObject(key, objectKey, inputStream, metadata);
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, File file) {
+  public StorageInfo putObject(String key, String objectKey, File file) {
     try (FileInputStream inputStream = new FileInputStream(file)) {
-      return this.putObject(bucketKey, objectKey, inputStream);
+      return this.putObject(key, objectKey, inputStream);
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, byte[] bytes, ObjectMetadata metadata) {
+  public StorageInfo putObject(String key, String objectKey, byte[] bytes, ObjectMetadata metadata) {
     try (ByteArrayInputStream inputStream = new ByteArrayInputStream(bytes)) {
-      return this.putObject(bucketKey, objectKey, inputStream, metadata);
+      return this.putObject(key, objectKey, inputStream, metadata);
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, byte[] bytes) {
+  public StorageInfo putObject(String key, String objectKey, byte[] bytes) {
     try (ByteArrayInputStream inputStream = new ByteArrayInputStream(bytes)) {
-      return this.putObject(bucketKey, objectKey, inputStream);
+      return this.putObject(key, objectKey, inputStream);
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, String url, ObjectMetadata metadata) {
+  public StorageInfo putObject(String key, String objectKey, String url, ObjectMetadata metadata) {
     try {
-      return this.putObject(bucketKey, objectKey, new URL(url).openStream(), metadata);
+      return this.putObject(key, objectKey, new URL(url).openStream(), metadata);
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, String objectKey, String url) {
+  public StorageInfo putObject(String key, String objectKey, String url) {
     try {
-      return this.putObject(bucketKey, objectKey, new URL(url).openStream());
+      return this.putObject(key, objectKey, new URL(url).openStream());
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult putObject(String bucketKey, PutObjectRequest putObjectRequest) {
+  public StorageInfo putObject(String key, PutObjectRequest putObjectRequest) {
     try {
-      Assert.notNull(bucketKey, "bucketKey must be not null");
+      Assert.notNull(key, "key must be not null");
       Assert.notNull(putObjectRequest, "putObjectRequest must be not null");
-      BucketProfile bp = ossProperties.getBucketProfile(bucketKey);
-      putObjectRequest.setBucketName(bp.getBucket());
-      log.info("putObject bucketKey={}, bucketName={}", bucketKey, bp.getBucket());
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      putObjectRequest.setBucketName(config.getBucket());
+      log.info("putObject key={}, bucketName={}", key, config.getBucket());
       log.info("putObject objectKey={}", putObjectRequest.getKey());
       PutObjectResult result = ossClient.putObject(putObjectRequest);
       Assert.notNull(result, "PutObjectResult must be not null");
       log.info("PutObjectResult {}", JSON.toJSONString(result));
-      StoreResult storeResult = new StoreResult();
-      storeResult.setBucket(bp.getBucket());
-      storeResult.setObjectKey(putObjectRequest.getKey());
-      storeResult.setObjectUrl(bp.getBaseUrl() + putObjectRequest.getKey());
-      storeResult.setEtag(result.getETag());
-      ossUploadCallback.callback(storeResult);
-      return storeResult;
+      StorageInfo storageInfo = new StorageInfo();
+      storageInfo.setBucket(config.getBucket());
+      storageInfo.setObjectKey(putObjectRequest.getKey());
+      storageInfo.setObjectUrl(config.getBaseUrl() + putObjectRequest.getKey());
+      storageInfo.setEtag(result.getETag());
+      ossUploadCallback.callback(storageInfo);
+      return storageInfo;
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, "文件上传失败：" + e.getMessage());
     }
   }
 
   @Override
-  public StoreResult copyObject(String fromBucketKey, String fromObjectKey, String toBucketKey, String toObjectKey, boolean isDelete) {
+  public StorageInfo copyObject(String fromKey, String fromObjectKey, String toKey, String toObjectKey, boolean isDelete) {
     try {
-      BucketProfile fromBucketProfile = ossProperties.getBucketProfile(fromBucketKey);
-      BucketProfile toBucketProfile = ossProperties.getBucketProfile(toBucketKey);
-      log.info("copyObject fromBucketKey={}, fromObjectKey={}", fromBucketKey, fromObjectKey);
-      log.info("copyObject toBucketKey={}, toObjectKey={}", toBucketKey, toObjectKey);
-      boolean doesObjectExist = ossClient.doesObjectExist(fromBucketProfile.getBucket(), fromObjectKey);
+      StorageConfig fromStorageConfig = ossProperties.getStorageConfig(fromKey);
+      StorageConfig toStorageConfig = ossProperties.getStorageConfig(toKey);
+      log.info("copyObject fromKey={}, fromObjectKey={}", fromKey, fromObjectKey);
+      log.info("copyObject toKey={}, toObjectKey={}", toKey, toObjectKey);
+      boolean doesObjectExist = ossClient.doesObjectExist(fromStorageConfig.getBucket(), fromObjectKey);
       ExceptionUtil.business(ErrorCodeDefaults.S0400, "源文件不存在：fromObjectKey=" + fromObjectKey).doThrow(!doesObjectExist);
-      CopyObjectResult result = ossClient.copyObject(fromBucketProfile.getBucket(), fromObjectKey, toBucketProfile.getBucket(), toObjectKey);
+      CopyObjectResult result = ossClient.copyObject(fromStorageConfig.getBucket(), fromObjectKey, toStorageConfig.getBucket(), toObjectKey);
       Assert.notNull(result, "CopyObjectResult must be not null");
       log.info("CopyObjectResult {}", JSON.toJSONString(result));
       // 复制成功，删除源文件
       if (isDelete) {
-        ossClient.deleteObject(fromBucketProfile.getBucket(), fromObjectKey);
+        ossClient.deleteObject(fromStorageConfig.getBucket(), fromObjectKey);
       }
-      StoreResult storeResult = new StoreResult();
-      storeResult.setBucket(toBucketProfile.getBucket());
-      storeResult.setObjectKey(toObjectKey);
-      storeResult.setObjectUrl(toBucketProfile.getBaseUrl() + toObjectKey);
-      storeResult.setEtag(result.getETag());
-      ossUploadCallback.callback(storeResult);
-      return storeResult;
+      StorageInfo storageInfo = new StorageInfo();
+      storageInfo.setBucket(toStorageConfig.getBucket());
+      storageInfo.setObjectKey(toObjectKey);
+      storageInfo.setObjectUrl(toStorageConfig.getBaseUrl() + toObjectKey);
+      storageInfo.setEtag(result.getETag());
+      ossUploadCallback.callback(storageInfo);
+      return storageInfo;
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, e.getMessage()).rewrite("复制文件失败");
     }
   }
 
   @Override
-  public OSSObject getObject(String bucketKey, String objectKey) {
+  public OSSObject getObject(String key, String objectKey) {
     try {
-      Assert.notNull(bucketKey, "bucketKey must be not null");
+      Assert.notNull(key, "key must be not null");
       Assert.notNull(objectKey, "objectKey must be not null");
-      BucketProfile bp = ossProperties.getBucketProfile(bucketKey);
-      log.info("OSSObject bucketKey={}, bucketName={}", bucketKey, bp.getBucket());
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      log.info("OSSObject key={}, bucketName={}", key, config.getBucket());
       log.info("OSSObject objectKey={}", objectKey);
-      OSSObject result = ossClient.getObject(bp.getBucket(), objectKey);
+      OSSObject result = ossClient.getObject(config.getBucket(), objectKey);
       Assert.notNull(result, "OSSObject must be not null");
       return result;
     } catch (Exception e) {
@@ -216,28 +216,28 @@ public class OssServiceImpl implements OssService {
   }
 
   @Override
-  public void deleteObject(String bucketKey, String objectKey) {
+  public void deleteObject(String key, String objectKey) {
     try {
-      Assert.notNull(bucketKey, "bucketKey must be not null");
+      Assert.notNull(key, "key must be not null");
       Assert.notNull(objectKey, "objectKey must be not null");
-      BucketProfile bp = ossProperties.getBucketProfile(bucketKey);
-      log.info("deleteObject bucketKey={}, bucketName={}", bucketKey, bp.getBucket());
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      log.info("deleteObject key={}, bucketName={}", key, config.getBucket());
       log.info("deleteObject objectKey={}", objectKey);
-      ossClient.deleteObject(bp.getBucket(), objectKey);
+      ossClient.deleteObject(config.getBucket(), objectKey);
     } catch (Exception e) {
       throw ExceptionUtil.business(ErrorCodeDefaults.A0700, e.getMessage()).rewrite("删除文件失败");
     }
   }
 
   @Override
-  public ObjectMetadata getObjectMetadata(String bucketKey, String objectKey) {
+  public ObjectMetadata getObjectMetadata(String key, String objectKey) {
     try {
-      Assert.notNull(bucketKey, "bucketKey must be not null");
+      Assert.notNull(key, "key must be not null");
       Assert.notNull(objectKey, "objectKey must be not null");
-      BucketProfile bp = ossProperties.getBucketProfile(bucketKey);
-      log.info("ObjectMetadata bucketKey={}, bucketName={}", bucketKey, bp.getBucket());
+      StorageConfig config = ossProperties.getStorageConfig(key);
+      log.info("ObjectMetadata key={}, bucketName={}", key, config.getBucket());
       log.info("ObjectMetadata objectKey={}", objectKey);
-      ObjectMetadata result = ossClient.getObjectMetadata(bp.getBucket(), objectKey);
+      ObjectMetadata result = ossClient.getObjectMetadata(config.getBucket(), objectKey);
       Assert.notNull(result, "ObjectMetadata must be not null");
       return result;
     } catch (Exception e) {