+++ Changelog.md
@@ -3,7 +3,7 @@
 - [x] [muyie-starter-apollo](muyie-starter-apollo)
 - [x] [muyie-starter-client](muyie-starter-client)
 - [x] [muyie-starter-mybatis](muyie-starter-mybatis)
-- [ ] [muyie-starter-oss](muyie-starter-oss)
+- [x] [muyie-starter-oss](muyie-starter-oss)
 - [ ] [muyie-starter-redis](muyie-starter-redis)
 - [ ] [muyie-starter-security](muyie-starter-security)
 - [x] [muyie-starter-thymeleaf](muyie-starter-thymeleaf)